+++ src/createStyles.ts
@@ -3,6 +3,7 @@ import merge from 'deepmerge';
 
 import { generateClassName } from './generateClassName.js';
 import { getPosthooks } from './plugins.js';
+import { SimpleStyleRegistry } from './simpleStyleRegistry.js';
 import { SimpleStyleRules } from './types.js';
 
 export type CreateStylesOptions = Partial<{
@@ -22,6 +23,15 @@ export type CreateStylesOptions = Partial<{
    * will render the injects <style /> before this element
    */
   insertBefore?: HTMLElement;
+
+  /**
+   * if set, will automatically prevent any styles from
+   * being flushed to the DOM or inserted at all.
+   * All styles will be accumulated in the registry
+   * and it will be up to you to determine how they should
+   * be flushed.
+   */
+  registry?: SimpleStyleRegistry;
 }>;
 
 function isNestedSelector(r: string): boolean {
@@ -44,6 +54,7 @@ function formatCSSRules(cssRules: Properties): string {
 }
 
 function execCreateStyles<T extends SimpleStyleRules, K extends keyof T, O extends Record<K, string>>(
+  ruleId: string,
   rules: T,
   options: CreateStylesOptions,
   parentSelector: string | null,
@@ -66,6 +77,7 @@ function execCreateStyles<T extends SimpleStyleRules, K extends keyof T, O exten
       guardCloseRuleWrite();
       mediaQueriesBuffer += `${classNameOrCSSRule}{`;
       const { mediaQueriesBuffer: mediaQueriesOutput, sheetBuffer: regularOutput } = execCreateStyles(
+        ruleId,
         classNameRules as T,
         options,
         parentSelector,
@@ -80,6 +92,7 @@ function execCreateStyles<T extends SimpleStyleRules, K extends keyof T, O exten
       const replaced = classNameOrCSSRule.replaceAll('&', parentSelector);
       for (const selector of replaced.split(/,\s*/)) {
         const { mediaQueriesBuffer: mediaQueriesOutput, sheetBuffer: regularOutput } = execCreateStyles(
+          ruleId,
           classNameRules as T,
           options,
           selector,
@@ -89,11 +102,12 @@ function execCreateStyles<T extends SimpleStyleRules, K extends keyof T, O exten
       }
     } else if (!parentSelector && typeof classNameRules === 'object') {
       guardCloseRuleWrite();
-      const generated = noGenerateClassName ? classNameOrCSSRule : generateClassName(classNameOrCSSRule);
+      const generated = noGenerateClassName ? classNameOrCSSRule : generateClassName(`${ruleId}_${classNameOrCSSRule}`);
       // @ts-expect-error - yes, we can index this object here, so be quiet
       out[classNameOrCSSRule] = generated;
       const generatedSelector = `${noGenerateClassName ? '' : '.'}${generated}`;
       const { mediaQueriesBuffer: mediaQueriesOutput, sheetBuffer: regularOutput } = execCreateStyles(
+        ruleId,
         classNameRules as T,
         options,
         generatedSelector,
@@ -133,19 +147,20 @@ function replaceBackReferences<O extends Record<string, string>>(out: O, sheetCo
   return getPosthooks().reduce((prev, hook) => hook(prev), outputSheetContents);
 }
 
-function createSheet(sheetContents: string) {
+function createSheet(ruleId: string, sheetContents: string) {
   const doc = globalThis.document as Partial<typeof globalThis.document> | null | undefined;
   if (doc === undefined) return null;
   // eslint-disable-next-line @typescript-eslint/no-deprecated
   if (typeof doc?.head?.appendChild !== 'function' || typeof doc.createElement !== 'function') return null;
   const styleTag = doc.createElement('style');
+  styleTag.id = ruleId;
   styleTag.innerHTML = sheetContents;
   return styleTag;
 }
 
-function flushSheetContents(sheetContents: string, options?: CreateStylesOptions) {
+function flushSheetContents(ruleId: string, sheetContents: string, options?: CreateStylesOptions) {
   // In case we're in come weird test environment that doesn't support JSDom
-  const styleTag = createSheet(sheetContents);
+  const styleTag = createSheet(ruleId, sheetContents);
   if (styleTag) {
     if (options?.insertAfter && options.insertBefore) {
       throw new Error('Both insertAfter and insertBefore were provided. Please choose only one.');
@@ -163,13 +178,14 @@ function coerceCreateStylesOptions(options?: CreateStylesOptions): CreateStylesO
   };
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function rawStyles<T extends SimpleStyleRules, K extends keyof T, O extends Record<K, string>>(
+export function rawStyles<T extends SimpleStyleRules>(
+  ruleId: string,
   rules: T,
   options?: Partial<CreateStylesOptions>,
 ) {
   const coerced = coerceCreateStylesOptions(options);
   const { sheetBuffer: sheetContents, mediaQueriesBuffer: mediaQueriesContents } = execCreateStyles(
+    ruleId,
     rules,
     coerced,
     null,
@@ -178,23 +194,46 @@ export function rawStyles<T extends SimpleStyleRules, K extends keyof T, O exten
 
   const mergedContents = `${sheetContents}${mediaQueriesContents}`;
 
-  if (coerced.flush) flushSheetContents(mergedContents, options);
+  if (coerced.flush) flushSheetContents(ruleId, mergedContents, options);
   return mergedContents;
 }
 
+export function makeRawStyles(registry: SimpleStyleRegistry) {
+  return function wrappedRawStyles<T extends SimpleStyleRules>(ruleId: string, rules: T) {
+    return rawStyles<T>(ruleId, rules, { registry });
+  };
+}
+
 export function keyframes<T extends Record<string, Properties>>(
+  ruleId: string,
   frames: T,
   options?: CreateStylesOptions,
 ): [string, string] {
   const coerced = coerceCreateStylesOptions(options);
-  const keyframeName = generateClassName('keyframes_');
-  const { sheetBuffer: keyframesContents } = execCreateStyles(frames, coerced, null, true);
+  const keyframeName = generateClassName(`${ruleId}_keyframes_`);
+  const { sheetBuffer: keyframesContents } = execCreateStyles(ruleId, frames, coerced, null, true);
   const sheetContents = `@keyframes ${keyframeName}{${keyframesContents}}`;
-  if (coerced.flush) flushSheetContents(sheetContents);
+  if (coerced.flush) flushSheetContents(ruleId, sheetContents);
   return [keyframeName, sheetContents];
 }
 
+export function makeKeyframes(registry: SimpleStyleRegistry) {
+  return function wrappedCreateKeyframes<T extends Record<string, Properties>>(ruleId: string, rules: T) {
+    return keyframes<T>(ruleId, rules, { registry });
+  };
+}
+
+export function makeCreateStyles(registry: SimpleStyleRegistry) {
+  return function wrappedCreateStyles<T extends SimpleStyleRules, K extends keyof T, O extends Record<K, string>>(
+    ruleId: string,
+    rules: T,
+  ) {
+    return createStyles<T, K, O>(ruleId, rules, { registry });
+  };
+}
+
 export default function createStyles<T extends SimpleStyleRules, K extends keyof T, O extends Record<K, string>>(
+  ruleId: string,
   rules: T,
   options?: Partial<CreateStylesOptions>,
 ) {
@@ -203,7 +242,7 @@ export default function createStyles<T extends SimpleStyleRules, K extends keyof
     classes: out,
     sheetBuffer: sheetContents,
     mediaQueriesBuffer: mediaQueriesContents,
-  } = execCreateStyles(rules, coerced, null);
+  } = execCreateStyles(ruleId, rules, coerced, null);
 
   const mergedContents = `${sheetContents}${mediaQueriesContents}`;
 
@@ -211,22 +250,22 @@ export default function createStyles<T extends SimpleStyleRules, K extends keyof
 
   let sheet: ReturnType<typeof flushSheetContents> = null;
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const updateSheet = <T2 extends SimpleStyleRules, K2 extends keyof T2, O2 extends Record<K2, string>>(
-    updatedRules: Partial<T2>,
-  ) => {
-    if ((options?.flush && sheet) || !options?.flush) {
+  const updateSheet = <T2 extends SimpleStyleRules>(updatedRules: Partial<T2>) => {
+    if (options?.flush || options?.registry || !options?.flush) {
       // We prefer the first set, and then we shallow merge
       const {
         classes: updatedOut,
         sheetBuffer: updatedSheetContents,
         mediaQueriesBuffer: updatedMediaQueriesContents,
-      } = execCreateStyles(merge(rules, updatedRules), { flush: false }, null);
+      } = execCreateStyles(ruleId, merge(rules, updatedRules), { flush: false }, null);
 
       const updatedMergedContents = `${updatedSheetContents}${updatedMediaQueriesContents}`;
 
       const updatedReplacedSheetContents = replaceBackReferences(out, updatedMergedContents);
       if (sheet) sheet.innerHTML = updatedReplacedSheetContents;
+      else if (options?.registry) {
+        options.registry.add(ruleId, updatedReplacedSheetContents);
+      }
       return { classes: updatedOut, stylesheet: updatedSheetContents } as {
         classes: typeof updatedOut;
         stylesheet: string;
@@ -235,7 +274,11 @@ export default function createStyles<T extends SimpleStyleRules, K extends keyof
     return null;
   };
 
-  if (coerced.flush) sheet = flushSheetContents(replacedSheetContents, options);
+  if (!options?.registry && coerced.flush) {
+    sheet = flushSheetContents(ruleId, replacedSheetContents, options);
+  } else if (options?.registry) {
+    options.registry.add(ruleId, replacedSheetContents);
+  }
   // Need this TS cast to get solid code assist from the consumption-side
   return {
     classes: out as unknown,